[SparseCompressed] support csc layout for add sparse/dense.
`add` when passed one sparse and one dense argument  will error if the
sparse argument does not have  csr layout. This PR modifies the
underlying algorithm to be generic on the compressed dimension handling
both csr and csc. The functions are renamed to use the
`sparse_compressed` qualifier rather than `sparse_csr`

Fixes: #114807

ghstack-source-id: 3c59fbc9be255874489184eb107b4a0cefb6b022
Pull Request resolved: https://github.com/pytorch/pytorch/pull/115433